+++ tests/regression/scalarfields/tests.py
@@ -4,8 +4,9 @@ from decimal import Decimal
 import stdnet
 from stdnet import test
 from stdnet.utils import populate, zip
+from stdnet.exceptions import FieldError
 
-from examples.models import TestDateModel, Statistics, Statistics2
+from examples.models import TestDateModel, Statistics, Statistics2, Page
 
 NUM_DATES = 100
 names = populate('string',NUM_DATES, min_len = 5, max_len = 20)
@@ -60,6 +61,38 @@ class TestAtomFields(test.TestCase):
         self.assertEqual(len(keys),1)
         self.assertEqual(keys[0],self.meta.autoid())
         
+
+class TestIntegerField(test.TestCase):
+    
+    def setUp(self):
+        self.orm.register(Page)
+    
+    def unregister(self):
+        self.orm.unregister(Page)
+        
+    def testDefaultValue(self):
+        p = Page()
+        self.assertEqual(p.in_navigation,1)
+        p = Page(in_navigation = '4')
+        self.assertEqual(p.in_navigation,'4')
+        p.save()
+        self.assertEqual(p.in_navigation,'4')
+        p = Page.objects.get(id = p.id)
+        self.assertEqual(p.in_navigation,4)
+        
+    def testNotValidated(self):
+        p = Page().save()
+        p = Page(in_navigation = 'bla')
+        self.assertRaises(FieldError,p.save)
+        
+    def testZeroValue(self):
+        p = Page(in_navigation = 0)
+        self.assertEqual(p.in_navigation,0)
+        p.save()
+        self.assertEqual(p.in_navigation,0)
+        p = Page.objects.get(id = p.id)
+        self.assertEqual(p.in_navigation,0)
+               
 
 class TestJsonField(test.TestCase):